Add .NET Core implementation of the InterestPaymentService project; add "UseSQLite" option; fix various outstanding issues. Make the project ready for running on Linux (Ubuntu 16.04)

+++ DepositCore/Models/NewDepositViewModels.cs
@@ -13,7 +13,7 @@ namespace DepositCore.Models
         [Display(Name = "Card number")]
         public string CardId { get; set; }
 
-        public byte WayOfAccumulationId { get; set; }
+        public int WayOfAccumulationId { get; set; }
 
         [Required]
         [RegularExpression("^[0-9]+([,][0-9]+)?$", ErrorMessage = "The field must contain only digits.")]
