+++ DomainLogic/IDepositStatesService.cs
@@ -2,7 +2,7 @@
 
 namespace DomainLogic
 {
-    public interface IDepositStatesService : IDbEntityService<DepositState, byte>
+    public interface IDepositStatesService : IDbEntityService<DepositState, int>
     {
         DepositState GetByName(string name);
     }
