+++ DomainLogic/Model/DepositTerm.cs
@@ -9,7 +9,7 @@ namespace DomainLogic.Model
         public int CurrencyId { get; set; }
 
         [Display(Name = "Months")]
-        public byte Months { get; set; }
+        public int Months { get; set; }
 
         [Display(Name = "Interest rate")]
         public decimal InterestRate { get; set; }
